Title: 문자에 따른 명령 2, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240121/문자에 따른 명령 2/text-based-commands2.cpp
@@ -4,38 +4,40 @@ using namespace std;
 
 //var
 
-int dy[4] = {-1,0,1,0};
-int dx[4] = {0,1,0,-1};
+int dy[4] = { -1,0,1,0 };
+int dx[4] = { 0,1,0,-1 };
 int dir_num;
 
-int x,y;
-int nx,ny;
+int x, y;
+int nx, ny;
 
-void init(){
-    x=0;y=0;
-    dir_num=0;
+void init() {
+    x = 0; y = 0;
+    dir_num = 0;
 
     string buffer;
-    cin>>buffer;
+    cin >> buffer;
 
-    int N;
+    long long N;
     char a;
     N = buffer.length();
-
-    for(int i=0;i<N;i++){
+    for (long long i = 0; i < N; i++) {
         a = buffer.at(i);
-        if(a == 'L'){// 왼쪽 90도 회전
-            dir_num = (dir_num - 1 +4 )%4;
-        }else if(a == 'R'){//오른쪽 90도 회전
-            dir_num = (dir_num + 1 +4 )%4;
-        }else if(a == 'F'){
+        if (a == 'L') {// 왼쪽 90도 회전
+            dir_num = (dir_num - 1 + 4) % 4;
+        }
+        else if (a == 'R') {//오른쪽 90도 회전
+            dir_num = (dir_num + 1 + 4) % 4;
+        }
+        else if (a == 'F') {
             x += dx[dir_num];
             y += dy[dir_num];
-        }else{}
+        }
+        else {}
     }
 
 
-    cout<<x<<" "<<y;
+    cout << x << " " << y;
 }
 
 int main() {
